feat: main 코드라인 구현
