+++ src/main/kotlin/lotto/Exchange.kt
@@ -1,0 +1,17 @@
+package lotto
+
+interface Exchange<Product> {
+    fun pay(money: Money)
+
+    fun product(): List<Product>
+
+    class Lotto(private val lottoDrawMachine: LottoDrawMachine) : Exchange<LottoNumber> {
+        private lateinit var lottoTicket: LottoTicket
+        override fun pay(money: Money) {
+            lottoTicket = LottoTicket(money)
+        }
+
+        override fun product(): List<LottoNumber> = (0 until lottoTicket.count)
+            .map { lottoDrawMachine.lottoNumber() }
+    }
+}
