+++ src/main/kotlin/lotto/LottoAgent.kt
@@ -1,0 +1,5 @@
+package lotto
+
+class LottoAgent(val exchange: Exchange<LottoNumber>) {
+    constructor(lottoExchange: LottoDrawMachine) : this(Exchange.Lotto(lottoExchange))
+}
